scripts/requirements.txt
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
docker==5.0.0
execnet==1.9.0
gitdb==4.0.7
GitPython==3.1.18
glob2==0.7
idna==2.10
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==2.0.1
packaging==21.0
parse==1.19.0
parse-type==0.5.2
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-bdd==4.1.0
pytest-forked==1.3.0
pytest-xdist==2.4.0
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
semantic-version==2.8.5
semver==2.13.0
six==1.16.0
smmap==4.0.0
toml==0.10.2
urllib3==1.26.5
websocket-client==1.2.1
analytics-python==1.4.0
pynacl==1.5.0
